Updates
Implemented an armour system. I probably did some other changes which i forgot to note, but please understand cuz i've taken like a 2h break twice while coding this. I also added an easter egg which if you can read code it will be easy to find, but know that it was made just for jokes and it should not be taken seriously!