New update
done some more work to implement boss fights, i could do it the same way i did a normal npc fight but that would be inefficient, i am trying to figure out a way to make stuff like the whole fight work with one code, and it just alternates depending on whether you are fighting a boss or a mini fight. I have a couple of ideas that include interfaces and make boss and enemy implement it, but we'll see about that in the near future.